refactor(courtboard): extract TennisCourtDisplay component
