+++ src/courtboard/components/TennisCourtDisplay.jsx
@@ -1,0 +1,542 @@
+import React, { useState, useEffect, useCallback } from 'react';
+import { CourtCard } from './CourtCard';
+import { WaitingList } from './WaitingList';
+import { NextAvailablePanel } from './NextAvailablePanel';
+
+// TennisBackend for real-time board subscription
+import { createBackend } from '../../registration/backend/index.js';
+const backend = createBackend();
+
+// Court availability helper - single source of truth for free/playable courts
+import { countPlayableCourts, listPlayableCourts } from '../../shared/courts/courtAvailability.js';
+
+// Import shared utilities from @lib
+import { TENNIS_CONFIG as _sharedTennisConfig } from '@lib';
+
+// ---- Debounce helper (no UI change) ----
+const debounce = (fn, ms = 150) => {
+  let t;
+  return (...a) => {
+    clearTimeout(t);
+    t = setTimeout(() => fn(...a), ms);
+  };
+};
+
+// ---- TENNIS_CONFIG: Override POLL_INTERVAL_MS for CourtBoard (faster updates) ----
+const TENNIS_CONFIG = {
+  ..._sharedTennisConfig,
+  TIMING: {
+    ..._sharedTennisConfig?.TIMING,
+    POLL_INTERVAL_MS: 2000,
+  },
+};
+
+/**
+ * TennisCourtDisplay - Main court display component
+ * Shows all 12 courts, waitlist, and availability panels
+ */
+export function TennisCourtDisplay() {
+  const isMobileView = window.IS_MOBILE_VIEW || false;
+  const [currentTime, setCurrentTime] = useState(new Date());
+  const [courts, setCourts] = useState(Array(12).fill(null));
+  const [waitlist, setWaitlist] = useState([]);
+  const [courtBlocks, setCourtBlocks] = useState([]); // Active blocks only (for availability)
+  const [upcomingBlocks, setUpcomingBlocks] = useState([]); // Future blocks today (for display)
+  const [checkStatusMinutes, setCheckStatusMinutes] = useState(150); // Default 150, loaded from settings
+  const [blockWarningMinutes, setBlockWarningMinutes] = useState(60); // Default 60, loaded from settings
+
+  // Mobile state - initialize from sessionStorage, updated via message from Mobile.html
+  const [mobileState, setMobileState] = useState(() => ({
+    registeredCourt: sessionStorage.getItem('mobile-registered-court'),
+    waitlistEntryId: sessionStorage.getItem('mobile-waitlist-entry-id'),
+  }));
+
+  // Listen for state updates from Mobile.html (MobileBridge broadcasts)
+  useEffect(() => {
+    const handleMessage = (event) => {
+      if (event.data?.type === 'mobile:state-updated') {
+        console.log('[Courtboard] Mobile state updated:', event.data.payload);
+        setMobileState(event.data.payload);
+      } else if (event.data?.type === 'refresh-board') {
+        // Triggered after waitlist:joined to check for waitlist-available notice
+        console.log('[Courtboard] Refresh board requested');
+        // The mobileState update from MobileBridge.broadcastState() will trigger
+        // the waitlist-available useEffect, but we can also manually trigger loadData
+        if (typeof window.refreshBoard === 'function') {
+          window.refreshBoard();
+        }
+      }
+    };
+
+    window.addEventListener('message', handleMessage);
+    return () => window.removeEventListener('message', handleMessage);
+  }, []);
+
+  // Time update
+  useEffect(() => {
+    const timer = setInterval(() => {
+      setCurrentTime(new Date());
+    }, 1000);
+    return () => clearInterval(timer);
+  }, []);
+
+  // Data loading - no longer reads from localStorage
+  // All state now comes from API via TennisBackend subscription
+  const loadData = useCallback(async () => {
+    // No-op: courts, waitlist, and courtBlocks now populate from API only
+    // Initial state is empty arrays, TennisBackend subscription fills them
+  }, []);
+
+  // DOM event listeners for cross-component updates
+  useEffect(() => {
+    // Legacy loadData call (now no-op, API subscription handles state)
+    loadData();
+
+    const handleUpdate = debounce(() => loadData(), 150);
+
+    window.addEventListener('tennisDataUpdate', handleUpdate, { passive: true });
+    window.addEventListener('DATA_UPDATED', handleUpdate, { passive: true });
+
+    // No polling - TennisBackend subscription provides real-time updates
+
+    return () => {
+      window.removeEventListener('tennisDataUpdate', handleUpdate);
+      window.removeEventListener('DATA_UPDATED', handleUpdate);
+    };
+  }, [loadData]);
+
+  // TennisBackend real-time subscription (primary data source)
+  useEffect(() => {
+    console.log('[Courtboard] Setting up TennisBackend subscription...');
+
+    const unsubscribe = backend.queries.subscribeToBoardChanges((domainBoard) => {
+      // Use pure Domain Board directly (legacy adapter removed)
+      const board = domainBoard;
+
+      console.log('[Courtboard] Board update received:', {
+        serverNow: board.serverNow,
+        courts: board.courts?.length,
+        waitlist: board.waitlist?.length,
+        upcomingBlocks: board.upcomingBlocks?.length,
+      });
+      console.log('[Courtboard Debug] Raw upcomingBlocks:', board.upcomingBlocks);
+
+      // Debug: log first 2 courts to see raw data
+      console.log('[Courtboard Debug] Raw board courts (first 2):', board.courts?.slice(0, 2));
+
+      // Update courts state
+      if (board.courts) {
+        // Transform API courts to Domain format for Courtboard rendering
+        // Domain format: court.session = { group: { players }, scheduledEndAt, startedAt }
+        const transformedCourts = Array(12)
+          .fill(null)
+          .map((_, idx) => {
+            const courtNumber = idx + 1;
+            const apiCourt = board.courts.find((c) => c && c.number === courtNumber);
+            if (!apiCourt) {
+              return null; // Empty court
+            }
+            if (!apiCourt.session && !apiCourt.block) {
+              return null; // No session or block
+            }
+
+            const players = (
+              apiCourt.session?.participants ||
+              apiCourt.session?.group?.players ||
+              []
+            ).map((p) => ({
+              name: p.displayName || p.name || 'Unknown',
+            }));
+
+            return {
+              session: apiCourt.session
+                ? {
+                    group: { players },
+                    scheduledEndAt: apiCourt.session.scheduledEndAt,
+                    startedAt: apiCourt.session.startedAt,
+                  }
+                : null,
+            };
+          });
+
+        // Debug: log first 2 transformed courts
+        console.log(
+          '[Courtboard Debug] Transformed courts (first 2):',
+          transformedCourts.slice(0, 2)
+        );
+        setCourts(transformedCourts);
+
+        // Extract active blocks from courts (for availability calculations)
+        const activeBlocks = board.courts
+          .filter((c) => c && c.block)
+          .map((c) => ({
+            id: c.block.id,
+            courtNumber: c.number,
+            reason: c.block.reason || c.block.title || 'Blocked',
+            startTime: c.block.startsAt,
+            endTime: c.block.endsAt,
+            isWetCourt: c.block.reason?.toLowerCase().includes('wet'),
+          }));
+        setCourtBlocks(activeBlocks);
+
+        // Extract upcoming blocks from API (future blocks for today, display only)
+        const futureBlocks = (board.upcomingBlocks || []).map((b) => ({
+          id: b.id,
+          courtNumber: b.courtNumber,
+          reason: b.title || b.reason || 'Blocked',
+          startTime: b.startTime,
+          endTime: b.endTime,
+          isWetCourt: (b.reason || b.title || '').toLowerCase().includes('wet'),
+        }));
+        setUpcomingBlocks(futureBlocks);
+      }
+
+      // Transform already-normalized waitlist from TennisQueries
+      // TennisQueries returns { group: { players } } format, we need { names } for rendering
+      const normalized = (board.waitlist || []).map((entry) => ({
+        id: entry.id,
+        position: entry.position,
+        groupType: entry.group?.type,
+        joinedAt: entry.joinedAt,
+        minutesWaiting: entry.minutesWaiting,
+        names: (entry.group?.players || []).map((p) => p.displayName || p.name || 'Unknown'),
+        players: entry.group?.players || [],
+      }));
+      console.log('[Courtboard] Transformed waitlist:', normalized);
+      setWaitlist(normalized);
+    });
+
+    console.log('[Courtboard] TennisBackend subscription active');
+
+    return () => {
+      console.log('[Courtboard] Unsubscribing from board updates');
+      unsubscribe();
+    };
+  }, []);
+
+  // Load check_status_minutes from system settings
+  useEffect(() => {
+    const loadSettings = async () => {
+      try {
+        console.log('[Courtboard] Loading settings, backend.admin:', !!backend.admin);
+        const result = await backend.admin?.getSettings?.();
+        console.log(
+          '[Courtboard] Settings result:',
+          result?.ok,
+          'check_status_minutes:',
+          result?.settings?.check_status_minutes
+        );
+        if (result?.ok && result.settings?.check_status_minutes) {
+          const minutes = parseInt(result.settings.check_status_minutes, 10);
+          if (minutes > 0) {
+            setCheckStatusMinutes(minutes);
+            console.log('[Courtboard] Loaded check_status_minutes:', minutes);
+          }
+        }
+        // Load block_warning_minutes
+        if (result?.ok && result.settings?.block_warning_minutes) {
+          const blockWarnMin = parseInt(result.settings.block_warning_minutes, 10);
+          if (blockWarnMin > 0) {
+            setBlockWarningMinutes(blockWarnMin);
+            console.log('[Courtboard] Loaded block_warning_minutes:', blockWarnMin);
+          }
+        }
+      } catch (err) {
+        console.warn('[Courtboard] Failed to load settings, using default:', err);
+      }
+    };
+    loadSettings();
+  }, []);
+
+  /**
+   * TWO-ROOT BRIDGE: Sync React state to window for mobile modal access.
+   *
+   * ARCHITECTURE INVARIANT:
+   * - This useEffect is the ONLY WRITER to window.CourtboardState
+   * - The mobile modal (second React root in MobileModalSheet) reads via getCourtboardState()
+   * - mobile-fallback-bar.js reads via getCourtboardState()
+   * - NO OTHER CODE should write to window.CourtboardState
+   *
+   * This bridge exists because the mobile modal is rendered in a separate React tree
+   * and cannot access this component's state directly.
+   */
+  useEffect(() => {
+    const now = new Date().toISOString();
+    const freeCount = countPlayableCourts(courts, courtBlocks, now);
+
+    console.log('[CourtboardState] Setting state:', {
+      courts: courts?.length,
+      courtBlocks: courtBlocks?.length,
+      upcomingBlocks: upcomingBlocks?.length,
+      waitingGroups: waitlist?.length,
+      freeCourts: freeCount,
+    });
+    window.CourtboardState = {
+      courts: courts,
+      courtBlocks: courtBlocks,
+      upcomingBlocks: upcomingBlocks,
+      waitingGroups: waitlist,
+      freeCourts: freeCount,
+      timestamp: Date.now(),
+    };
+
+    // Update mobile button state after state is set
+    if (typeof window.updateJoinButtonState === 'function') {
+      console.log('[CourtboardState] Calling updateJoinButtonState');
+      window.updateJoinButtonState();
+    } else {
+      console.log('[CourtboardState] updateJoinButtonState not found');
+    }
+  }, [courts, courtBlocks, upcomingBlocks, waitlist]);
+
+  // Auto-show waitlist-available notice when court is free and THIS mobile user is first in waitlist
+  useEffect(() => {
+    if (!isMobileView) return;
+
+    const hasWaitlist = waitlist.length > 0;
+    if (!hasWaitlist) {
+      // No waitlist - close notice if open
+      if (window.MobileModal?.currentType === 'waitlist-available') {
+        window.MobileModal?.close?.();
+      }
+      return;
+    }
+
+    // Check if THIS mobile user is first in the waitlist
+    // Use mobileState (React state) instead of sessionStorage for reactivity
+    const mobileWaitlistEntryId = mobileState.waitlistEntryId;
+    const firstGroup = waitlist[0];
+    const isUserFirstInWaitlist = mobileWaitlistEntryId && firstGroup?.id === mobileWaitlistEntryId;
+
+    // Use shared helper for consistent free court calculation
+    const now = new Date().toISOString();
+    const freeCourtCount = countPlayableCourts(courts, courtBlocks, now);
+    const freeCourtList = listPlayableCourts(courts, courtBlocks, now);
+
+    console.log('[WaitlistNotice] Check:', {
+      freeCourts: freeCourtCount,
+      freeCourtList: freeCourtList,
+      waitlistLength: waitlist?.length,
+      isMobileView: isMobileView,
+      mobileWaitlistEntryId: mobileWaitlistEntryId,
+      firstGroupId: firstGroup?.id,
+      isUserFirstInWaitlist: isUserFirstInWaitlist,
+      shouldShow: freeCourtCount > 0 && isUserFirstInWaitlist,
+      totalCourts: courts?.length,
+      courtsWithSession: courts?.filter((c) => c?.session).length,
+    });
+
+    if (freeCourtCount > 0 && isUserFirstInWaitlist) {
+      // Court available AND this mobile user is first in waitlist - show notice
+      window.MobileModal?.open('waitlist-available', { firstGroup });
+    } else if (window.MobileModal?.currentType === 'waitlist-available') {
+      // Not first, no free courts, or no waitlist - close notice if it's currently showing
+      window.MobileModal?.close?.();
+    }
+  }, [courts, courtBlocks, waitlist, isMobileView, mobileState]);
+
+  window.refreshBoard = loadData;
+
+  // Build status map using courts state from API (no localStorage fallback)
+  let statusByCourt = {};
+  let selectableByCourt = {};
+  let statusObjectByCourt = {};
+  // Build data object from React courts state for status computation
+  const data = {
+    courts: courts,
+    waitlist: waitlist.map((g) => ({
+      id: g.id,
+      players: g.names.map((n) => ({ name: n })),
+    })),
+  };
+
+  try {
+    const A = window.Tennis?.Domain?.availability || window.Tennis?.Domain?.Availability;
+    if (A) {
+      const now = new Date();
+      // Use courtBlocks from React state instead of localStorage
+      const blocks = courtBlocks || [];
+
+      const wetSet = new Set(
+        (blocks || [])
+          .filter(
+            (b) =>
+              b?.isWetCourt &&
+              new Date(b.startTime ?? b.start) <= now &&
+              now < new Date(b.endTime ?? b.end)
+          )
+          .map((b) => b.courtNumber)
+      );
+      const _statuses = A.getCourtStatuses({ data, now, blocks, wetSet }) || [];
+      statusByCourt = Object.fromEntries(_statuses.map((s) => [s.courtNumber, s.status]));
+      selectableByCourt = Object.fromEntries(_statuses.map((s) => [s.courtNumber, s.selectable]));
+      statusObjectByCourt = Object.fromEntries(_statuses.map((s) => [s.courtNumber, s]));
+    }
+  } catch (e) {
+    console.warn('Error building status map:', e);
+  }
+
+  const hasWaiting = waitlist.length > 0;
+
+  return (
+    <div className="h-screen min-h-screen bg-gradient-to-br from-slate-700 to-slate-600 p-4 text-white flex flex-col">
+      {/* Header */}
+      <div className="mb-4">
+        <div className="w-full flex justify-center">
+          <div className="flex items-center relative">
+            {hasWaiting && (
+              <div className="absolute left-[-140px] lg:left-[-100px] xl:left-[-140px]">
+                <svg
+                  width="40"
+                  height="40"
+                  className="lg:w-20 lg:h-20 xl:w-32 xl:h-32"
+                  viewBox="0 0 24 24"
+                  fill="none"
+                  xmlns="http://www.w3.org/2000/svg"
+                >
+                  <path d="M5 20V4" stroke="#D4D4D8" strokeWidth="1.5" strokeLinecap="round" />
+                  <path d="M5 4L17 7L5 10V4Z" fill="#FB923C" stroke="#EA580C" strokeWidth="1" />
+                </svg>
+              </div>
+            )}
+            <div className="text-center">
+              <div
+                className="time-header font-light text-white mb-1"
+                style={{
+                  fontFamily:
+                    'system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif',
+                }}
+              >
+                {currentTime.toLocaleTimeString([], { hour: 'numeric', minute: '2-digit' })}
+              </div>
+              <div
+                className="date-header text-gray-300 -mt-1"
+                style={{
+                  fontFamily:
+                    'system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif',
+                }}
+              >
+                {currentTime.toLocaleDateString([], {
+                  weekday: 'long',
+                  month: 'long',
+                  day: 'numeric',
+                })}
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+
+      {/* Main Content */}
+      <div className="flex-1 main-layout h-full min-h-0 items-stretch">
+        {/* LEFT column wrapper */}
+        <div className="left-column courts-section h-full flex flex-col min-h-0" data-left-col>
+          <div className="bg-slate-800/50 rounded-xl shadow-2xl p-4 h-full backdrop-blur flex flex-col min-h-0">
+            {/* Desktop layout - Top row */}
+            <div className="courts-grid mb-4">
+              {[1, 2, 3, 4, 5, 6, 7, 8].map((num) => (
+                <div key={num} className="flex justify-center">
+                  <CourtCard
+                    courtNumber={num}
+                    currentTime={currentTime}
+                    statusByCourt={statusByCourt}
+                    selectableByCourt={selectableByCourt}
+                    statusObjectByCourt={statusObjectByCourt}
+                    data={data}
+                    isMobileView={isMobileView}
+                    checkStatusMinutes={checkStatusMinutes}
+                    upcomingBlocks={upcomingBlocks}
+                    blockWarningMinutes={blockWarningMinutes}
+                    courts={courts}
+                    courtBlocks={courtBlocks}
+                  />
+                </div>
+              ))}
+            </div>
+            <div className="h-3 bg-gray-400 mb-4 rounded-full flex-shrink-0 divider-line" />
+
+            {/* Bottom section */}
+            <div className="bottom-section min-h-0 overflow-auto">
+              <WaitingList
+                waitlist={waitlist}
+                courts={courts}
+                currentTime={currentTime}
+                courtBlocks={courtBlocks}
+                upcomingBlocks={upcomingBlocks}
+                maxWaitingDisplay={TENNIS_CONFIG.DISPLAY?.MAX_WAITING_DISPLAY || 4}
+              />
+              <div className="courts-grid-bottom">
+                {[12, 11, 10, 9].map((num) => (
+                  <div key={num} className="flex justify-center">
+                    <CourtCard
+                      courtNumber={num}
+                      currentTime={currentTime}
+                      statusByCourt={statusByCourt}
+                      selectableByCourt={selectableByCourt}
+                      statusObjectByCourt={statusObjectByCourt}
+                      data={data}
+                      isMobileView={isMobileView}
+                      checkStatusMinutes={checkStatusMinutes}
+                      upcomingBlocks={upcomingBlocks}
+                      blockWarningMinutes={blockWarningMinutes}
+                      courts={courts}
+                      courtBlocks={courtBlocks}
+                    />
+                  </div>
+                ))}
+              </div>
+            </div>
+
+            {/* Mobile layout */}
+            <div className="mobile-layout">
+              <div className="mobile-courts-grid">
+                {[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12].map((num) => (
+                  <CourtCard
+                    key={num}
+                    courtNumber={num}
+                    currentTime={currentTime}
+                    statusByCourt={statusByCourt}
+                    selectableByCourt={selectableByCourt}
+                    statusObjectByCourt={statusObjectByCourt}
+                    data={data}
+                    isMobileView={isMobileView}
+                    checkStatusMinutes={checkStatusMinutes}
+                    upcomingBlocks={upcomingBlocks}
+                    blockWarningMinutes={blockWarningMinutes}
+                    courts={courts}
+                    courtBlocks={courtBlocks}
+                  />
+                ))}
+              </div>
+
+              {hasWaiting && (
+                <div className="mobile-waiting-section">
+                  <WaitingList
+                    waitlist={waitlist}
+                    courts={courts}
+                    currentTime={currentTime}
+                    courtBlocks={courtBlocks}
+                    upcomingBlocks={upcomingBlocks}
+                    maxWaitingDisplay={TENNIS_CONFIG.DISPLAY?.MAX_WAITING_DISPLAY || 4}
+                  />
+                </div>
+              )}
+            </div>
+          </div>
+        </div>
+
+        {/* RIGHT column wrapper */}
+        <div className="right-column flex flex-col h-full min-h-0" data-right-col>
+          <NextAvailablePanel
+            courts={courts}
+            currentTime={currentTime}
+            waitlist={waitlist}
+            courtBlocks={courtBlocks}
+            upcomingBlocks={upcomingBlocks}
+            maxDisplay={_sharedTennisConfig?.DISPLAY?.MAX_WAITING_DISPLAY || 6}
+          />
+        </div>
+      </div>
+    </div>
+  );
+}
